Refresh customer table working
I was able to use a pyqt table widget to store customers email, name and create a checkbox to select if they would like there invoice printed. I made function in QBservices.py that will always return a list of dictionarys with all customers and there information. Got the last refreshed label to update when refresh list is clicked.