Sender working
I created a customer class that has a few functions for adding and updating customers to a class variable. When clicking on refresh customer list it clears the list and creates new customers using the list of dictionaries from QBservices. Also it sets prt to False by default. When clicking on a check box it will update the entire customer class list everytime (i dont know any other way to do it).

If any customer is Checked in the customer list, your default web browser will open the pdf file that was selected. If the customer was not checked, an outlook draft will open with the subject and to lines populated with the correct customer belonging to the invoice. Also the invoice will be attached.