finished working on sender - for now
The sender is functioning as intended, however it requires having outlook installed on the computer. I'd like it to use whatever the computers default mail app is. That'll be something to work on in the next update.

I worked on the logic for the Send button. When clicking send a few conditions will determine if the invoice will print or email and weather the email body will have a table with invoice details or a custom body input by the user.